Updated the description.
Fixed undefined index and variables.

Added support for the Roles and Permissions add-on (need to test this!).

Changed the confirmation drop down to a checkbox, need to figure the placement for this, or possibly remove it. Not sure what will work best in this situation.

Made sure the price options are updated correctly when moving to a new event. It should use the default price option of the new event.

TODO:
Update the prices when the attendee is moved to a new event. This should use the default price of the new event.

[touch:2007]